Fleshed out YOLO layer, but haven't tested it out yet.
TODO: Test out with image

Eager execution allows the output of the YOLO head to display the
tensors. However, I haven't figured out how to make the YOLO layer
as part of the output to `keras.Model`.

A workaround is to channel the outputs into the YOLO layer, though
this solution is not very clean.